CRIMAP-219 namespace V1 operations and models
A v2 api is being added. V2 will use postgres and not DynamoDB.
This change namespaces V1 code in preparation for that change.
