+++ app/services/operations/dynamodb/get_application.rb
@@ -1,0 +1,15 @@
+module Operations
+  module Dynamodb
+    class GetApplication
+      attr_reader :application_id
+
+      def initialize(application_id)
+        @application_id = application_id
+      end
+
+      def call
+        ::Dynamodb::CrimeApplication.find(application_id)
+      end
+    end
+  end
+end
