+++ spec/services/operations/dynamodb/get_application_spec.rb
@@ -1,18 +1,18 @@
 require 'rails_helper'
 
-RSpec.describe Operations::GetApplication do
+RSpec.describe Operations::Dynamodb::GetApplication do
   subject { described_class.new(application_id) }
 
   let(:application_id) { '12345' }
 
   let(:query) { { id: application_id } }
-  let(:document) { instance_double(CrimeApplication) }
+  let(:document) { instance_double(described_class) }
 
   describe '#call' do
     context 'when application exists' do
       before do
         allow(
-          CrimeApplication
+          ::Dynamodb::CrimeApplication
         ).to receive(:where).with(query).and_return([document])
       end
